float struggle continues

+++ libft/ft_pow.c
@@ -6,15 +6,15 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/10 10:53:47 by mviinika          #+#    #+#             */
-/*   Updated: 2022/03/24 19:45:37 by mviinika         ###   ########.fr       */
+/*   Updated: 2022/03/25 16:09:15 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long double	ft_pow(long double num, long double pow)
+double	ft_pow(double num, double pow)
 {
-	long double	res;
+	double	res;
 	int		counter;
 
 	res = 1;
